Maya exporter automatically picks the mesh and root bone off the selected object. Interface extended for animation management. Added support for file signature handling to the SerializeManager to detect invalid file input.

+++ src/engine/geLoad3ds.cpp
@@ -299,7 +299,7 @@ namespace GE
     //Try to open file
     File module = File::GetModule();
     file = module.getRelativeFile( filename );
-    fileSize = file.getSize();
+    fileSize = (int) file.getSize();
     if( !file.open( "rb" )) return false;
 
     //Read main chunk info
